API Reference

MediaMixModellingDemo.generate_fourier_series — Function

generate_fourier_series(t, p=365.25, n=5)

Generates fourier series with period p and degree n (the higher, the more flexible it is) It can be then fitted with coefficients to mimic any period trend

Expects t to be a time index series

Returns array of shape: (size(t,1),2n)

Example

seaso=generatefourierseries(1:400,365.25, 5)

MediaMixModellingDemo.generate_seasonality_features — Method

generate_seasonality_features(t, p=365.25, n=5)

Generates seasonality features given an array of tuples in a format (period,degree) Eg, 7-day period of degree 3 would be

Expects t to be a time index series

Returns array of shape: (size(t,1),2n)

Example

seaso=generatefourierseries(1:400,365.25, 5)

MediaMixModellingDemo.geometric_decay — Method

geometric_decay(x::Vector{T},decay_rate,normalize=true) where {T<:Real}

Simple geometric decay transformation if normalize=true it divides the output by the sum of the geometric series Note: Does NOT check if decay_rate<1 etc to ensure that the sum convergences to the analytic formula

MediaMixModellingDemo.getflatsamples — Method

getflatsamples(chain,groupname)

Extract a group of variables under name groupname from chain and flattens all samples into the first dimension ie, outputs a dimension: (numsamples*numchains,num_variables)

MediaMixModellingDemo.plot_periodogram — Function

function plot_periodogram(input_arr)

Plot Fourier transform coefficients to uncover the most prominent frequencies / seasonalities Assumes equally spaced data points Looks only for periods that have seen at least 2 full cycles (ie, size ÷ 2 at maximum!) Shows top-k values

Example

p=10 # period is 10 y=sin.(2π/p*collect(1:20)) # generate 20 data points plot_periodogram(y,1) # plot periodogram, period=10 should be highlighted as maximum

MediaMixModellingDemo.standardize_by_max — Method

standardize_by_max(X)

Max()-only transform to allow easy scaling between features and the outcome Uses MinMax() pipe under the hood but overwrites the minimum to be =0

Example: y_std,pipe_cache_y=standardize_by_max(select(df,target_label))

MediaMixModellingDemo.standardize_by_zscore — Method

standardize_by_zscore(X)

Zscore transform to center the feature to its mean value and scale it (to make it easier to set priors)

Example: y_std,pipe_cache_y=standardize_by_zscore(select(df,target_label))