API Reference

Index

MediaMixModellingDemo.ParamsPlot

MediaMixModellingDemo.ParamsStage1

MediaMixModellingDemo.ParamsStage2

MediaMixModellingDemo.build_inputs

MediaMixModellingDemo.calc_mroas

MediaMixModellingDemo.calc_roas

MediaMixModellingDemo.calc_roas

MediaMixModellingDemo.convert_budget_multiplier_to_spend_multiplier

MediaMixModellingDemo.fit

MediaMixModellingDemo.fit

MediaMixModellingDemo.generate_fourier_series

MediaMixModellingDemo.generate_objective_func

MediaMixModellingDemo.generate_objective_func

MediaMixModellingDemo.generate_seasonality_features

MediaMixModellingDemo.geometric_decay

MediaMixModellingDemo.getflatsamples

MediaMixModellingDemo.hill_curve

MediaMixModellingDemo.hill_curve

MediaMixModellingDemo.mean_fitted_effects

MediaMixModellingDemo.multiply_and_normalize

MediaMixModellingDemo.optimize

MediaMixModellingDemo.plot_contributions

MediaMixModellingDemo.plot_effects_vs_spend

MediaMixModellingDemo.plot_mmm_one_pager

MediaMixModellingDemo.plot_model_fit_by_period

MediaMixModellingDemo.plot_optimization_one_pager

MediaMixModellingDemo.plot_optimized_contribution

MediaMixModellingDemo.plot_optimized_spend_share_comparison

MediaMixModellingDemo.plot_optimized_uplift_histogram

MediaMixModellingDemo.plot_periodogram

MediaMixModellingDemo.plot_priors_decay_rate

MediaMixModellingDemo.plot_response_curves_table

MediaMixModellingDemo.quick_nuts_diagnostics

MediaMixModellingDemo.saturate_adspend

MediaMixModellingDemo.set_priors

MediaMixModellingDemo.set_priors

MediaMixModellingDemo.set_priors_auto_scales

MediaMixModellingDemo.set_priors_stage2_decay_rates

MediaMixModellingDemo.simulate_revenues_summed

MediaMixModellingDemo.standardize_by_max

MediaMixModellingDemo.standardize_by_zscore

MediaMixModellingDemo.sum_columns

MediaMixModellingDemo.threaded_objective_func

MediaMixModellingDemo.to_masked_matrix

MediaMixModellingDemo.workflow_budget_to_simulation

RecipesBase.plot

RecipesBase.plot

## Docstrings

MediaMixModellingDemo.ParamsPlot — Type

struct ParamsPlot
    title_suffix = ""
    color_spend = palette(:default)[6]
    color_spend_optimized = palette(:default)[6]
    color_revenues = palette(:default)[1]
    color_revenues_original = palette(:default)[2]
    units_revenues = "USD"
    units_spend = "USD"
    title_fontsize = 10
    table_fontsize_header = 10
    table_fontsize_body = 8
    output_dpi = 150
    output_size_mmm = (1000, 800)
    output_size_optim = (1000, 800)
end

Holds plotting defaults

MediaMixModellingDemo.ParamsStage1 — Type

struct ParamsStage1
    model_name=Val(:any)
    scales_trend_offset::Float64=0.
    scales_growth_trend::Float64=1.
    scales_trend::Float64 = 0.2
    scales_hols::Float64 = 0.3
    scales_seas::AbstractArray{Float64} = ones(Float64,1)
    scales_feat::AbstractArray{Float64} = ones(Float64,1)
    scales_noise::Float64 = 0.2
    cat_levels::Int = 1
end

Holds priors and relevant parameters for Stage 1 of the modelling

Arguments:

model_name=Val(:any) : Symbol representing the model version being fitted

scales_trend_offset::Float64=0. : Scale ("width") of the Normal RV alpha (intercepts for different groups)

scales_growth_trend::Float64=1. : Scale ("width") of the Normal RV growth_trend (trend component)

scales_trend::Float64 = 0.2 : Scale ("width") of the Normal RV beta_trend for input X_trend (flexible trend fitting, eg, with splines)

scales_hols::Float64 = 0.3 : Scale ("width") of the Normal RV beta_hols for input X_hols (holidays features)

scales_seas::AbstractArray{Float64} = ones(Float64,1) : Scale ("width") of the Normal RV beta_seas for input X_feat (seasonality components)

scales_feat::AbstractArray{Float64} = ones(Float64,1) : Scale ("width") of the Normal RV beta_feat for input X_feat (concatenated features)

scales_noise::Float64 = 0.2 : Extent of the random noise around deterministic trend / rate parameter of Exponential distribution for sigma

cat_levels::Int = 1 : Number of categorical levels in RV alpha that will be provided in X_cat vector (allows for different intercepts for different groups)

Example:

Set priors with automated priors like this:

p1=ParamsStage1(
    model_name=Val(Symbol(MODEL_NAME_PREFIT)),
    scales_trend=0.2,
    scales_hols=0.3,
    scales_noise=0.2,
    cat_levels=1
)
p1=set_priors_stage1_trendline(y_std,p1)
p1=set_priors_auto_scales(y_std,X_seas,:scales_seas,1.0,p1)
p1=set_priors_auto_scales(y_std,X_feat,:scales_feat,1.0,p1)

MediaMixModellingDemo.ParamsStage2 — Type

struct ParamsStage2
    model_name=Val(:any)
    scales_trend_offset::Float64 = 0.3
    scales_trend::Float64 = 0.2
    scales_noise::Float64 = 0.3
    scales_context::AbstractArray{Float64} = ones(Float64,1)
    scales_org::AbstractArray{Float64} = ones(Float64,1)

    decay_rate_alphas::AbstractArray{Float64} = ones(Float64,1)
    decay_rate_betas::AbstractArray{Float64} = ones(Float64,1)

    adspend_mean_nonzero::AbstractArray{Float64} = ones(Float64,1)
    adspend_median::AbstractArray{Float64} = ones(Float64,1)

    locs_spend_halfmaxpoint::AbstractArray{Float64} = ones(Float64,1)
    scales_spend_halfmaxpoint::AbstractArray{Float64} = ones(Float64,1)

    locs_spend_beta::AbstractArray{Float64} = ones(Float64,1)
    scales_spend_beta::AbstractArray{Float64} = ones(Float64,1)
    units_ratio_spend_to_y::AbstractArray{Float64} = ones(Float64,1)
    factor_to_roas_of_one::AbstractArray{Float64} = units_ratio_spend_to_y .* 2
end

Holds priors and relevant parameters for Stage 2 of the modelling

For easy set up, use utility functions that build it from the inputs

Arguments:

model_name=Val(:any) : Symbol representing the model version being fitted

scales_trend_offset::Float64=0. : Scale ("width") of the Normal RV alpha (intercepts for different groups)

scales_trend::Float64 = 0.2 : Scale ("width") of the Normal RV beta_trend for input X_trend (flexible trend fitting, eg, with splines)

scales_noise::Float64 = 0.2 : Extent of the random noise around deterministic trend / rate parameter of Exponential distribution for sigma

scales_context::AbstractArray{Float64} = ones(Float64,1) : Scale ("width") of the Normal RV beta_context for input X_context (context variables)

scales_org::AbstractArray{Float64} = ones(Float64,1) : Scale ("width") of the Normal RV beta_org for input X_feat (organic variables - can have only POSITIVE effect)

decay_rate_alphas::AbstractArray{Float64} = ones(Float64,1) : decay_rate RV is modelled by Beta distribution, alpha is the corresponding parameter

decay_rate_betas::AbstractArray{Float64} = ones(Float64,1) : decay_rate RV is modelled by Beta distribution, beta is the corresponding parameter

adspend_mean_nonzero::AbstractArray{Float64} = ones(Float64,1) : Calculated quantity of the average non-zero spend (used to initialize halfmaxpoint)

adspend_median::AbstractArray{Float64} = ones(Float64,1) : Calculated quantity of median of the spend

locs_spend_halfmaxpoint::AbstractArray{Float64} = ones(Float64,1) : Center of the Normal RV halfmaxpoint for the halfmax concentration point in Hill Curve (can be initiated by average of the non-zero spend)

scales_spend_halfmaxpoint::AbstractArray{Float64} = ones(Float64,1) : Scale ("width") of the Normal RV halfmaxpoint for the halfmax concentration point in Hill Curve

locs_spend_beta::AbstractArray{Float64} = ones(Float64,1) : Center of the Normal RV 'beta_spend' that represents the ROAS when the ad spend is at halfmaxpoint

scales_spend_beta::AbstractArray{Float64} = ones(Float64,1) : Scale ("width") of the Normal RV 'beta_spend' that represents the ROAS when the ad spend is at halfmaxpoint

units_ratio_spend_to_y::AbstractArray{Float64} = ones(Float64,1) : Ratio of ad spend to Y to be able to convert unit effect (used for factor_to_roas_of_one)

factor_to_roas_of_one::AbstractArray{Float64} : Conversion factor that ensures that provided beta_spend represents the ROAS when the ad spend is at halfmaxpoint

Example:

p2=ParamsStage2(
    model_name=Val(Symbol(MODEL_NAME)),
    scales_trend_offset=0.3,
    scales_trend=0.2,
    scales_noise=0.3,
)

p2 = set_priors_auto_scales(y_std,X_context,:scales_context,1.0,p2)
p2 = set_priors_auto_scales(y_std,X_org,:scales_org,1.0,p2)
p2 = set_priors_stage2_hill_curves(X_spend,p2;
        units_ratio_spend_to_y=getindex.(pipe_cache_spend,:xh)/pipe_cache_y[1].xh,
        halfmaxpoint_scale=0.3,expected_roas=1.0, expected_roas_scale=1.5)
p2 = set_priors_stage2_decay_rates(["ooh","digital","digital","ooh","ooh"],decay_rates_types_dictionary,p2,cols_spend)

sanity_check_priors(p2;X_spend,X_context,X_org);

MediaMixModellingDemo.build_inputs — Method

build_inputs(df::AbstractDataFrame;
    cols_spend::AbstractVector{String},
    col_target::String="revenue",
    col_datetime::String="dt",
    col_time_std::String="time_std",
    col_cat::String="",
    cols_hols::AbstractVector{String}=String[],
    cols_context::AbstractVector{String}=String[],
    cols_organic::AbstractVector{String}=String[],
    seasonality_periods::AbstractVector{Float64}=Float64[],
    spline_degree::Int=0,
    fit_stage2_mask::AbstractVector{Bool}=trues(nrow(df)),
    optim_mask::AbstractVector{Bool}=trues(nrow(df))
)

Builds the input data object for the successing fitting & optimization

Arguments:

col_target::String="revenue" - column name of the target/response variable (eg, revenues)

col_datetime::String="dt" - column name of the variable with dates

col_time_std::String="time_std" - column name of the standardized time index (from 0-1, strict)

col_cat::String="events" - column name of the categorical variable representing various events (ie not available); Set to "" if there isn't any

cols_context::AbstractVector{String}=String[] - column name of the context variables (eg, macroeconomic indicators, market trends, competitors sales or promotions) (variables will be standardized via Z-score); Defaults to empty if there are none

cols_organic::AbstractVector{String}=String[] - column name of the organic marketing activities (eg, email newsletters) (variables will be scaled to maximum=1 via Max() function and their effect on response can be only positive); Defaults to empty if there are none

cols_hols::AbstractVector{String}=String[] - column name of the holiday indicators; Defaults to empty if there are none

cols_spend::AbstractVector{String} - column names of the Ad spend variables that we want to model (variables will be scaled to maximum=1 via Max() function)

seasonality_periods::AbstractVector{Float64}=Float64[] - what seasonalities are expected in the trendline; Defaults to empty if there are none

spline_degree::Int=0 - if complicated trend modelling is needed, what degree of spline basis should be used (uses cubic bases splines from Splines2.jl)

fit_stage2_mask::AbstractVector{Bool}=trues(nrow(df)) - mask to be apllied to observed data in Stage 2 of modelling (use the observation when =true); Defaults to all observations being used

optim_mask::AbstractVector{Bool}=trues(nrow(df)) - mask to be apllied in the budget optimization (use the observation when =true); Defaults to all observations being used

MediaMixModellingDemo.calc_mroas — Method

calc_mroas(x::Number, delta::Number, chain::AbstractMCMC.AbstractChains, p, slice_idx::Int)

Calculate marginal ROAS (mROAS) at a given point x with a delta (=step size) for a variable under slice_idx (ie, 3rd channel would have slice_idx=3)

Example: Calculate mROAS for all Ad spend variables with delta=0.01

# p2=ParamsStage2() # model parameters from stage 2
# chain is the result of fitting of stage 2 model
# cols_spend are ad spend column names
mroas_at_mean=[calc_mroas(p2.adspend_mean_nonzero[idx],0.01,chain,p2,idx)[1] for idx in 1:length(cols_spend)]

MediaMixModellingDemo.calc_roas — Function

calc_roas(effect::Number, spend::Number, factor_spend_to_effect::Number=1)

Calculates ROAS with an option to normalize the effect to the right scale (if Y and X are scaled)

MediaMixModellingDemo.calc_roas — Method

calc_roas(effects::AbstractArray, spends::AbstractArray, factors_spend_to_effect::AbstractArray,weights::AbstractArray)

Calculates Total ROAS (of all Ad channels) as a weighted-average of individual ROAS' weighted by the raw spend

MediaMixModellingDemo.convert_budget_multiplier_to_spend_multiplier — Method

convert_budget_multiplier_to_spend_multiplier(spend_prev_trf::AbstractVector,
                                                   factor_to_scale_spend_to_orig::AbstractVector,
                                                   budget_multiplier::AbstractVector)

Converts a relative budget_multiplier (a vector of %s for each marketing channel, eg, [1,1.2,0.8]) to a spend_multiplier (=output), which retains the total amount of spend in the original domain Ie, money money between channel while not changing the total spend

Example:

spend_prev_trf=X_spend|>Matrix|>sum_columns|>vec
factor_to_scale_spend_to_orig=getindex.(pipe_cache_spend,:xh)
budget_multiplier=ones(size(X_spend,2))
spend_multiplier=convert_budget_multiplier_to_spend_multiplier(spend_prev_trf,factor_to_scale_spend_to_orig,budget_multiplier)

MediaMixModellingDemo.fit — Function

fit(inputs::InputData,fitted_stage1::Stage1Fit,params::ParamsStage2,model_func::Function,
    algorithm::Turing.Inference.AdaptiveHamiltonian=NUTS(300,0.65;max_depth=10);
    mcmc_samples=250,mcmc_chains=4)

Fits Stage 2 model (marketing transformations of interest) based on model (model_func) with provided data (inputs) and priors (params) fitted_stage1 is a Fit from Stage 1 that holds the trend components (see Documentation for the rationale behind 2-stage fit) Defaults to NUTS algorithm

MediaMixModellingDemo.fit — Function

fit(inputs::InputData,params::ParamsStage1,model_func::Function,
    algorithm::Turing.Inference.AdaptiveHamiltonian=NUTS(300,0.65;max_depth=10);
    mcmc_samples=100,mcmc_chains=1)

Fits Stage 1 model (trend components) based on model (model_func) with provided data (inputs) and priors (params) Defaults to NUTS algorithm

MediaMixModellingDemo.generate_fourier_series — Function

generate_fourier_series(t, p=365.25, n=5)

Generates fourier series with period p and degree n (the higher, the more flexible it is) It can be then fitted with coefficients to mimic any period trend

Expects t to be a time index series

Returns array of shape: (size(t,1),2n)

Example:

seasonality_generate_fourier_series(1:400,365.25, 5)

MediaMixModellingDemo.generate_objective_func — Function

generate_objective_func(
    chain_optim::AbstractMCMC.AbstractChains,
        model_orig::DynamicPPL.Model,
    X_spend::AbstractMatrix, optim_mask::BitVector,
    spend_raw_sum::AbstractVector,
    loss_func::Function = identity;
    simulations_basecase = nothing,
    extract_key::Symbol=:y)

Objective function generator which runs directly off a budget_multiplier operating on the ad spend in original domain and producing uplift in terms of revenues (or a chosen extract_key) Auxilary objective (gx) minimizes the delta between original adspend and new adspend (it must be strictly smaller or equal)

Notes:

This version leaves ad spend budget balancing to the optimization algorithm

The simplest loss_func function that you can use is identity!

Optional: simulation_basecase can be provided to speed up the optimization (ie, avoid re-computing the revenue under the old budget)

If optim_mask subsets data (ie, !=trues(size(Xspend,1))) then only the underlying segment of the `Xspend` matrix is changed and only that portion of simulated revenues is considered for objective function

Example:

# Prepare inputs
chain_optim=Chains(chain,:parameters)
simulations_prev=simulate_revenues_summed(chain_optim,model_orig,optim_mask;extract_key=:mu)

# boundaries on possible solution
lower_bound = 0.5*ones(length(cols_spend)) # max 50% reduction
upper_bound = 1.5*ones(length(cols_spend)) # max 50% increase
bounds = [lower_bound upper_bound]'

# Bayesian Decision Theory -- how to weigh the outcomes across the posterior distribution
# define a simple asymmetric (risk-averse) loss function
loss_func(x)=x>0 ? 0.5x : x

# All channels must have some spend in the optimization period!
@assert all((@view(X_spend[optim_mask,:])|>sum_columns) .>0)

# Method with direct budget multiplier
# spend_raw_sum is masked with optim_mask!
spend_raw_sum=revert_pipe_spend(X_spend[optim_mask,:])|>sum_columns

objective_func=generate_objective_func(
    chain_optim,model_orig,Matrix(X_spend),optim_mask,
    spend_raw_sum,loss_func;simulations_basecase=simulations_prev,extract_key=:mu)

MediaMixModellingDemo.generate_objective_func — Method

generate_objective_func(chain_optim::AbstractMCMC.AbstractChains,
model_orig::DynamicPPL.Model,
X_spend::AbstractMatrix, optim_mask::BitVector,
factor_to_scale_spend_to_orig::AbstractVector,
loss_func::Function,
bounds::AbstractVecOrMat;
simulations_basecase = nothing,
extract_key::Symbol=:y)

Generates objective function that depends only on budget_multiplier input and returns simulated revenues and other requirements of Metaheuristics Checks that implied spend_multiplier_new (calculated from budget_multiplier) is within the provided bounds as auxilary objective

Notes:

This version is suitable if optimization algorithm struggles to find solutions that maintain the same adspend (it balances it under the hood)

The simplest loss_func function that you can use is identity!

Optional: simulation_basecase can be provided to speed up the optimization (ie, avoid re-computing the revenue under the old budget)

If optim_mask subsets data (ie, !=trues(size(Xspend,1))) then only the underlying segment of the `Xspend` matrix is changed and only that portion of simulated revenues is considered for objective function

MediaMixModellingDemo.generate_seasonality_features — Method

generate_seasonality_features(t, p=365.25, n=5)

Generates seasonality features given an array of tuples in a format (period,degree) Eg, 7-day period of degree 3 would be

Expects t to be a time index series

Returns array of shape: (size(t,1),2n)

Example:

seasonality_=generate_fourier_series(1:400,365.25, 5)

MediaMixModellingDemo.geometric_decay — Method

geometric_decay(x::Vector{T},decay_rate,normalize=true) where {T<:Real}

Simple geometric decay transformation if normalize=true it divides the output by the sum of the geometric series Note: Does NOT check if decay_rate<1 etc to ensure that the sum convergences to the analytic formula

MediaMixModellingDemo.getflatsamples — Method

getflatsamples(chain,groupname)

Extract a group of variables under name groupname from chain and flattens all samples into the first dimension ie, outputs a dimension: (numsamples*numchains,num_variables)

MediaMixModellingDemo.hill_curve — Method

hill_curve(x, half_max_concentration_point, hill_coef,::Val{:safe})

Hill Curve-based spend saturation. Safe implementation that works better with ForwardDiff

MediaMixModellingDemo.hill_curve — Method

hill_curve(x, half_max_concentration_point, hill_coef)

Hill Curve-based spend saturation based on: https://www.physiologyweb.com/calculators/hillequationinteractive_graph.html

MediaMixModellingDemo.mean_fitted_effects — Method

extract_fitted_effects(::Val{:model_stage2a},generated_quant,
    extract_keys=[:mu_trend,:mu_org,:mu_context,:mu_spend_by_var],mask=nothing)

Extract specific keys from the generated_quantities produced by Turing generated_quantities() and concatenates them It should be implemented for each model to ensure the right logic

Returns: a Vector (!)

Example: mean_fitted_effects(Val(:model_stage2a),stage2_fit_allsamples) mean_fitted_effects(Val(:model_stage2a),stage2_fit_allsamples;extract_keys=[:mu_spend_by_var])

MediaMixModellingDemo.multiply_and_normalize — Method

multiply_and_normalize(v::AbstractVector, multipliers::AbstractVector)

Multiplies a vector v (of adspend) by multipliers in a way that doesn't change the overall sum

MediaMixModellingDemo.optimize — Method

optimize(fitted::Stage2Fit, inputs::InputData;
              metaheuristics_options::Metaheuristics.Options = nothing,
              multiplier_bounds = nothing)

Runs optimization loop and returns results in the struct OptimalBudget

MediaMixModellingDemo.plot_contributions — Method

plot_contributions(effect_shares, cols, p)

Creates a waterfall plot to show % contributions to fitted revenues (no-noise! only deterministic components)

MediaMixModellingDemo.plot_effects_vs_spend — Method

plot_effects_vs_spend(effect_shares, spend_shares, cols, p)

Plots a comparison between where we spend money (spend_shares) vs effect on revenues (effect_shares). Expressed in relative terms (% of total)

Big differences within each category (spend % vs effect %) suggest optimization opportunities

MediaMixModellingDemo.plot_mmm_one_pager — Method

plot_mmm_one_pager(plot_array, frame_idx::Int64, p)

Combines all plots in plot array into 1 layout for MMM 1-pager

Allows to reveal the plots partially, one by one (provide frame_idx)

MediaMixModellingDemo.plot_model_fit_by_period — Method

plot_model_fit_by_period(y_true, y_pred, p)

Create a plot comparing the modelled revenues(y_pred) vs actuals (y_true)

MediaMixModellingDemo.plot_optimization_one_pager — Method

plot_optimization_one_pager(plot_array, frame_idx::Int64, p)

Create a layout for Optimization 1-pager given a provided array of plots (plot_array)

It is possible to reveal plots one by one (use frame_idx)

MediaMixModellingDemo.plot_optimized_contribution — Method

plot_optimized_contribution(effect_prev, effect_optim, roas_total, optim_start,
                                 optim_end, revert_y_func, p)

Plots expected delta in revenue if the new marketing budget has been implemented (on the same period)

MediaMixModellingDemo.plot_optimized_spend_share_comparison — Method

plot_optimized_spend_share_comparison(spend_share_prev, spend_share_optim, cols, p)

Plots a comparison between previous marketing spend allocation and the optimum discovered

Note that the differences are constrainted by the optimization settings (see multiplier_bounds !)

MediaMixModellingDemo.plot_optimized_uplift_histogram — Method

plot_optimized_uplift_histogram(revenue_uplift, p)

Plots a histogram of modelled uplift (revenue_uplift)

MediaMixModellingDemo.plot_periodogram — Function

function plot_periodogram(input_arr,top_k)

Plot Fourier transform coefficients to uncover the most prominent frequencies / seasonalities Assumes equally spaced data points Looks only for periods that have seen at least 2 full cycles (ie, size ÷ 2 at maximum!) Shows top-k values

Example

p=10 # period is 10
y=sin.(2π/p*collect(1:20)) # generate 20 data points
plot_periodogram(y,1) # plot periodogram, period=10 should be highlighted as maximum

MediaMixModellingDemo.plot_priors_decay_rate — Function

plot_priors_decay_rate(p,cols=nothing)

Plots priors for decay rate in terms of Beta distributions

Example:

p2=ParamsStage2()
cols_spend=nothing # available from data prep
plot_priors_decay_rate(p2,cols_spend)

MediaMixModellingDemo.plot_response_curves_table — Method

plot_response_curves_table(decay_rates, roass, mroas_at_means, cols, roas_total, p)

Plot a quick summary table of the parameters of the response curves

Expects vectors of parameters - one of each channel

MediaMixModellingDemo.quick_nuts_diagnostics — Method

quick_nuts_diagnostics(chain::AbstractMCMC.AbstractChains, max_depth::Int)

Prints quick diagnostics of NUTS algorithm - mostly importantly alerting the user if there have been any divergences

MediaMixModellingDemo.saturate_adspend — Method

saturate_adspend(x::Number, chain::AbstractMCMC.AbstractChains, factor_to_roas_of_one::AbstractVector)

Extracts Hill Curve parameters and the corresponding beta coefficients from provided chain and applies them to a provided point (x)

WARNING: Depends on model implementation using the same RVs as modelstage2a, ie, `betaspend,halfmaxpoint,slope`

MediaMixModellingDemo.set_priors — Method

set_priors(params::ParamsStage1,inputs::InputData)

Sets automated priors based on input data for Stage 1 fitting

MediaMixModellingDemo.set_priors — Method

set_priors(params::ParamsStage2,inputs::InputData)

Sets automated priors based on input data for Stage 2 fitting

MediaMixModellingDemo.set_priors_auto_scales — Function

set_priors_auto_scales(y::AbstractArray,X,param_key::Symbol,factor::Union{Number,AbstractArray}=1.0,
    existing_params=ParamsStage1())

Sets scales such that 1 standard dev of input data (X) could trigger at most 3standard dev. response of Y (y) (because beta coefficient are Normal). This can be increased by factor (defaults to 1.0x). Result is saved into a Parameters object (as per constructor) under a key param_key It can extend (selectively update) existing set of parameters (provide to argument existing_params)

Special behaviour:

Values are clipped to range 0.05 - 10. (warning will be issued if it's exceeded)

Shortcuts to straight passthrough if X=nothing

MediaMixModellingDemo.set_priors_stage2_decay_rates — Function

set_priors_stage2_decay_rates(decay_rate_types::AbstractArray{String},
    decay_rates_types_dictionary::Dict=decay_rates_types_dictionary,
    existing_params=ParamsStage2(),var_names=nothing)

Sets decay rates priors (for Beta distribution) as a dictionary lookup of different options available in decay_rates_types_dictionary (eg, "tv", "digital")

Example:

types_=["ooh","digital","digital","ooh","ooh"]
set_priors_stage2_decay_rates(types_,decay_rates_types_dictionary,p2,cols_spend)

MediaMixModellingDemo.simulate_revenues_summed — Method

simulate_revenues_summed(chain_optim::AbstractMCMC.AbstractChains, model_optim::DynamicPPL.Model,

optimmask::BitVector;extractkey::Symbol=:y)

Samples posterior predictive from model model_optim conditioned on Chains chain_optim. Optional - you can provide a mask optim_mask (for time dimensions/1st dimension) and Symbol for the posterior predictive value (extract_key)

Example:

simulations_prev=simulate_revenues_summed(chain_optim,model_orig,optim_mask;extract_key=:y)

For budget simulations we replace inputs in the model like:

chain_optim=Chains(chain,:parameters)
model_args_prev=model_orig.args;
# replace the old spend with new
model_args_new=merge(model_args_prev,(;X_spend=X_spend_new));
simulations=simulate_revenues_summed(chain_optim,model_stage2a(model_args_new...),optim_mask);

MediaMixModellingDemo.standardize_by_max — Method

standardize_by_max(X)

Max()-only transform to allow easy scaling between features and the outcome Uses MinMax() pipe under the hood but overwrites the minimum to be =0

Example:

y_std,pipe_cache_y=standardize_by_max(select(df,target_label))

MediaMixModellingDemo.standardize_by_zscore — Method

standardize_by_zscore(X)

Zscore transform to center the feature to its mean value and scale it (to make it easier to set priors)

Example:

y_std,pipe_cache_y=standardize_by_zscore(select(df,target_label))

MediaMixModellingDemo.sum_columns — Method

sum_columns(x::AbstractMatrix)

Unified interface that sums columns of a provided Matrix/DataFrame/Vector Returns a Vector (!)

MediaMixModellingDemo.threaded_objective_func — Method

threaded_objective_func(budget_multipliers,objective_func)

Parallelizes provided function objective_func for Metaheuristics optimization loop by leveraging available threads in your Julia instance

Example:

Make sure to change the keyword to parallel_evaluation=true

options = Metaheuristics.Options(time_limit=10.,debug=false,parallel_evaluation=true)
@time result = Metaheuristics.optimize(x->threaded_objective_func(x,objective_func), bounds,
    Metaheuristics.ECA(N=7*2*length(cols_spend),K=7,η_max=2.,options=options))

MediaMixModellingDemo.to_masked_matrix — Function

to_masked_matrix(x::DataFrame,mask=trues(size(x,1)))

Converts DataFrame to a matrix and applies a mask to its rows if provided

MediaMixModellingDemo.workflow_budget_to_simulation — Method

workflow_budget_to_simulation(
                chain_optim::AbstractMCMC.AbstractChains, model_orig::DynamicPPL.Model,
               X_spend::AbstractMatrix, optim_mask::BitVector,
               factor_to_scale_spend_to_orig::AbstractVector,
               budget_multiplier::AbstractVector)

For a given context (budget_multiplier) produces new adspend (X_spend_new) and associated simulated revenues (smulations_new) Used to produced an objective function for the optimization that depends only on budget_multiplier

RecipesBase.plot — Function

Plots.plot(fitted::Stage2Fit, inputs::InputData, pplot::ParamsPlot = ParamsPlot())

Wraps all computations to create MMM 1-Pager

RecipesBase.plot — Function

Plots.plot(optimal::OptimalBudget, fitted::Stage2Fit, inputs::InputData,
                pplot::ParamsPlot = ParamsPlot())

Wraps all computations required to plot the Optimization 1-pager