API Reference

Index

MediaMixModellingDemo.ParamsStage1

MediaMixModellingDemo.ParamsStage2

MediaMixModellingDemo.calc_mroas

MediaMixModellingDemo.calc_roas

MediaMixModellingDemo.calc_roas

MediaMixModellingDemo.convert_budget_multiplier_to_spend_multiplier

MediaMixModellingDemo.generate_fourier_series

MediaMixModellingDemo.generate_objective_func

MediaMixModellingDemo.generate_objective_func

MediaMixModellingDemo.generate_seasonality_features

MediaMixModellingDemo.geometric_decay

MediaMixModellingDemo.getflatsamples

MediaMixModellingDemo.hill_curve

MediaMixModellingDemo.hill_curve

MediaMixModellingDemo.mean_fitted_effects

MediaMixModellingDemo.multiply_and_normalize

MediaMixModellingDemo.plot_periodogram

MediaMixModellingDemo.plot_priors_decay_rate

MediaMixModellingDemo.saturate_adspend

MediaMixModellingDemo.set_priors_auto_scales

MediaMixModellingDemo.set_priors_stage2_decay_rates

MediaMixModellingDemo.simulate_revenues_summed

MediaMixModellingDemo.standardize_by_max

MediaMixModellingDemo.standardize_by_zscore

MediaMixModellingDemo.sum_columns

MediaMixModellingDemo.threaded_objective_func

MediaMixModellingDemo.to_masked_matrix

MediaMixModellingDemo.workflow_budget_to_simulation

## Docstrings

MediaMixModellingDemo.ParamsStage1 — Type

struct ParamsStage1
    model_name=Val(:any)
    scales_trend_offset::Float64=0. 
    scales_growth_trend::Float64=1.
    scales_trend::Float64 = 0.2
    scales_hols::Float64 = 0.3
    scales_seas::AbstractArray{Float64} = ones(Float64,1)
    scales_feat::AbstractArray{Float64} = ones(Float64,1)
    scales_noise::Float64 = 0.2 
    cat_levels::Int = 1 
end

Holds priors and relevant parameters for Stage 1 of the modelling

Arguments:

model_name=Val(:any) : Symbol representing the model version being fitted

scalestrendoffset::Float64=0. : Scale ("width") of the Normal RV alpha (intercepts for different groups)

scalesgrowthtrend::Float64=1. : Scale ("width") of the Normal RV growth_trend (trend component)

scalestrend::Float64 = 0.2 : Scale ("width") of the Normal RV `betatrend` for input X_trend (flexible trend fitting, eg, with splines)

scaleshols::Float64 = 0.3 : Scale ("width") of the Normal RV `betahols` for input X_hols (holidays features)

scalesseas::AbstractArray{Float64} = ones(Float64,1) : Scale ("width") of the Normal RV `betaseas` for input X_feat (seasonality components)

scalesfeat::AbstractArray{Float64} = ones(Float64,1) : Scale ("width") of the Normal RV `betafeat` for input X_feat (concatenated features)

scales_noise::Float64 = 0.2 : Extent of the random noise around deterministic trend / rate parameter of Exponential distribution for sigma

catlevels::Int = 1 : Number of categorical levels in RV alpha that will be provided in Xcat vector (allows for different intercepts for different groups)

Example:

Set priors with automated priors like this:

p1=ParamsStage1(
    model_name=Val(Symbol(MODEL_NAME_PREFIT)),
    scales_trend=0.2,
    scales_hols=0.3,
    scales_noise=0.2,
    cat_levels=1
)
p1=set_priors_stage1_trendline(y_std,p1)
p1=set_priors_auto_scales(y_std,X_seas,:scales_seas,1.0,p1)
p1=set_priors_auto_scales(y_std,X_feat,:scales_feat,1.0,p1)

MediaMixModellingDemo.ParamsStage2 — Type

struct ParamsStage2
    model_name=Val(:any)
    scales_trend_offset::Float64 = 0.3
    scales_trend::Float64 = 0.2
    scales_noise::Float64 = 0.3
    scales_context::AbstractArray{Float64} = ones(Float64,1)
    scales_org::AbstractArray{Float64} = ones(Float64,1)

    decay_rate_alphas::AbstractArray{Float64} = ones(Float64,1)
    decay_rate_betas::AbstractArray{Float64} = ones(Float64,1)

    adspend_mean_nonzero::AbstractArray{Float64} = ones(Float64,1)
    adspend_median::AbstractArray{Float64} = ones(Float64,1)

    locs_spend_halfmaxpoint::AbstractArray{Float64} = ones(Float64,1)
    scales_spend_halfmaxpoint::AbstractArray{Float64} = ones(Float64,1)

    locs_spend_beta::AbstractArray{Float64} = ones(Float64,1) 
    scales_spend_beta::AbstractArray{Float64} = ones(Float64,1)
    units_ratio_spend_to_y::AbstractArray{Float64} = ones(Float64,1)
    factor_to_roas_of_one::AbstractArray{Float64} = units_ratio_spend_to_y .* 2
end

Holds priors and relevant parameters for Stage 2 of the modelling

For easy set up, use utility functions that build it from the inputs

Arguments:

model_name=Val(:any) : Symbol representing the model version being fitted

scalestrendoffset::Float64=0. : Scale ("width") of the Normal RV alpha (intercepts for different groups)

scalestrend::Float64 = 0.2 : Scale ("width") of the Normal RV `betatrend` for input X_trend (flexible trend fitting, eg, with splines)

scales_noise::Float64 = 0.2 : Extent of the random noise around deterministic trend / rate parameter of Exponential distribution for sigma

scalescontext::AbstractArray{Float64} = ones(Float64,1) : Scale ("width") of the Normal RV `betacontext` for input X_context (context variables)

scalesorg::AbstractArray{Float64} = ones(Float64,1) : Scale ("width") of the Normal RV `betaorg` for input X_feat (organic variables - can have only POSITIVE effect)

decayratealphas::AbstractArray{Float64} = ones(Float64,1) : decay_rate RV is modelled by Beta distribution, alpha is the corresponding parameter

decayratebetas::AbstractArray{Float64} = ones(Float64,1) : decay_rate RV is modelled by Beta distribution, beta is the corresponding parameter

adspendmeannonzero::AbstractArray{Float64} = ones(Float64,1) : Calculated quantity of the average non-zero spend (used to initialize halfmaxpoint)

adspend_median::AbstractArray{Float64} = ones(Float64,1) : Calculated quantity of median of the spend

locsspendhalfmaxpoint::AbstractArray{Float64} = ones(Float64,1) : Center of the Normal RV halfmaxpoint for the halfmax concentration point in Hill Curve (can be initiated by average of the non-zero spend)

scalesspendhalfmaxpoint::AbstractArray{Float64} = ones(Float64,1) : Scale ("width") of the Normal RV halfmaxpoint for the halfmax concentration point in Hill Curve

locsspendbeta::AbstractArray{Float64} = ones(Float64,1) : Center of the Normal RV 'beta_spend' that represents the ROAS when the ad spend is at halfmaxpoint

scalesspendbeta::AbstractArray{Float64} = ones(Float64,1) : Scale ("width") of the Normal RV 'beta_spend' that represents the ROAS when the ad spend is at halfmaxpoint

unitsratiospendtoy::AbstractArray{Float64} = ones(Float64,1) : Ratio of ad spend to Y to be able to convert unit effect (used for factor_to_roas_of_one)

factortoroasofone::AbstractArray{Float64} : Conversion factor that ensures that provided beta_spend represents the ROAS when the ad spend is at halfmaxpoint

Example:

p2=ParamsStage2(
    model_name=Val(Symbol(MODEL_NAME)),
    scales_trend_offset=0.3,
    scales_trend=0.2,
    scales_noise=0.3,
)

p2 = set_priors_auto_scales(y_std,X_context,:scales_context,1.0,p2)
p2 = set_priors_auto_scales(y_std,X_org,:scales_org,1.0,p2)
p2 = set_priors_stage2_hill_curves(X_spend,p2;
        units_ratio_spend_to_y=getindex.(pipe_cache_spend,:xh)/pipe_cache_y[1].xh,
        halfmaxpoint_scale=0.3,expected_roas=1.0, expected_roas_scale=1.5)
p2 = set_priors_stage2_decay_rates(["ooh","digital","digital","ooh","ooh"],decay_rates_types_dictionary,p2,cols_spend)

sanity_check_priors(p2;X_spend,X_context,X_org);

MediaMixModellingDemo.calc_mroas — Method

calc_mroas(x::Number, delta::Number, chain::AbstractMCMC.AbstractChains, p, slice_idx::Int)

Calculate marginal ROAS (mROAS) at a given point x with a delta (=step size) for a variable under slice_idx (ie, 3rd channel would have slice_idx=3)

Example: Calculate mROAS for all Ad spend variables with delta=0.01

# p2=ParamsStage2() # model parameters from stage 2
# chain is the result of fitting of stage 2 model
# cols_spend are ad spend column names
mroas_at_mean=[calc_mroas(p2.adspend_mean_nonzero[idx],0.01,chain,p2,idx)[1] for idx in 1:length(cols_spend)]

MediaMixModellingDemo.calc_roas — Function

calc_roas(effect::Number, spend::Number, factor_spend_to_effect::Number=1)

Calculates ROAS with an option to normalize the effect to the right scale (if Y and X are scaled)

MediaMixModellingDemo.calc_roas — Method

calc_roas(effects::AbstractArray, spends::AbstractArray, factors_spend_to_effect::AbstractArray,weights::AbstractArray)

Calculates Total ROAS (of all Ad channels) as a weighted-average of individual ROAS' weighted by the raw spend

MediaMixModellingDemo.convert_budget_multiplier_to_spend_multiplier — Method

convert_budget_multiplier_to_spend_multiplier(spend_prev_trf::AbstractVector,
                                                   factor_to_scale_spend_to_orig::AbstractVector,
                                                   budget_multiplier::AbstractVector)

Converts a relative budget_multiplier (a vector of %s for each marketing channel, eg, [1,1.2,0.8]) to a spend_multiplier (=output), which retains the total amount of spend in the original domain Ie, money money between channel while not changing the total spend

Example:

spend_prev_trf=X_spend|>Matrix|>sum_columns|>vec
factor_to_scale_spend_to_orig=getindex.(pipe_cache_spend,:xh)
budget_multiplier=ones(size(X_spend,2))
spend_multiplier=convert_budget_multiplier_to_spend_multiplier(spend_prev_trf,factor_to_scale_spend_to_orig,budget_multiplier)

MediaMixModellingDemo.generate_fourier_series — Function

generate_fourier_series(t, p=365.25, n=5)

Generates fourier series with period p and degree n (the higher, the more flexible it is) It can be then fitted with coefficients to mimic any period trend

Expects t to be a time index series

Returns array of shape: (size(t,1),2n)

Example

seaso=generatefourierseries(1:400,365.25, 5)

MediaMixModellingDemo.generate_objective_func — Function

generate_objective_func(
    chain_optim::AbstractMCMC.AbstractChains, 
        model_orig::DynamicPPL.Model,
    X_spend::AbstractMatrix, optim_mask::BitVector,
    spend_raw_sum::AbstractVector,
    loss_func::Function = identity; 
    simulations_basecase = nothing,
    extract_key::Symbol=:y)

Objective function generator which runs directly off a budget_multiplier operating on the ad spend in original domain and producing uplift in terms of revenues (or a chosen extract_key) Auxilary objective (gx) minimizes the delta between original adspend and new adspend (it must be strictly smaller or equal)

Notes:

This version leaves ad spend budget balancing to the optimization algorithm

The simplest loss_func function that you can use is identity!

Optional: simulation_basecase can be provided to speed up the optimization (ie, avoid re-computing the revenue under the old budget)

If optim_mask subsets data (ie, !=trues(size(Xspend,1))) then only the underlying segment of the `Xspend` matrix is changed and only that portion of simulated revenues is considered for objective function

Example:

# Prepare inputs
chain_optim=Chains(chain,:parameters)
simulations_prev=simulate_revenues_summed(chain_optim,model_orig,optim_mask;extract_key=:mu)

# boundaries on possible solution
lower_bound = 0.5*ones(length(cols_spend)) # max 50% reduction
upper_bound = 1.5*ones(length(cols_spend)) # max 50% increase
bounds = [lower_bound upper_bound]'

# Bayesian Decision Theory -- how to weigh the outcomes across the posterior distribution
# define a simple asymmetric (risk-averse) loss function
loss_func(x)=x>0 ? 0.5x : x

# All channels must have some spend in the optimization period!
@assert all((@view(X_spend[optim_mask,:])|>sum_columns) .>0)

# Method with direct budget multiplier
# spend_raw_sum is masked with optim_mask!
spend_raw_sum=revert_pipe_spend(X_spend[optim_mask,:])|>sum_columns

objective_func=generate_objective_func(
    chain_optim,model_orig,Matrix(X_spend),optim_mask,
    spend_raw_sum,loss_func;simulations_basecase=simulations_prev,extract_key=:mu)

MediaMixModellingDemo.generate_objective_func — Method

generate_objective_func(chain_optim::AbstractMCMC.AbstractChains, 
model_orig::DynamicPPL.Model,
X_spend::AbstractMatrix, optim_mask::BitVector,
factor_to_scale_spend_to_orig::AbstractVector,
loss_func::Function,
bounds::AbstractVecOrMat;
simulations_basecase = nothing,
extract_key::Symbol=:y)

Generates objective function that depends only on budget_multiplier input and returns simulated revenues and other requirements of Metaheuristics Checks that implied spend_multiplier_new (calculated from budget_multiplier) is within the provided bounds as auxilary objective

Notes:

This version is suitable if optimization algorithm struggles to find solutions that maintain the same adspend (it balances it under the hood)

The simplest loss_func function that you can use is identity!

Optional: simulation_basecase can be provided to speed up the optimization (ie, avoid re-computing the revenue under the old budget)

If optim_mask subsets data (ie, !=trues(size(Xspend,1))) then only the underlying segment of the `Xspend` matrix is changed and only that portion of simulated revenues is considered for objective function

MediaMixModellingDemo.generate_seasonality_features — Method

generate_seasonality_features(t, p=365.25, n=5)

Generates seasonality features given an array of tuples in a format (period,degree) Eg, 7-day period of degree 3 would be

Expects t to be a time index series

Returns array of shape: (size(t,1),2n)

Example

seaso=generatefourierseries(1:400,365.25, 5)

MediaMixModellingDemo.geometric_decay — Method

geometric_decay(x::Vector{T},decay_rate,normalize=true) where {T<:Real}

Simple geometric decay transformation if normalize=true it divides the output by the sum of the geometric series Note: Does NOT check if decay_rate<1 etc to ensure that the sum convergences to the analytic formula

MediaMixModellingDemo.getflatsamples — Method

getflatsamples(chain,groupname)

Extract a group of variables under name groupname from chain and flattens all samples into the first dimension ie, outputs a dimension: (numsamples*numchains,num_variables)

MediaMixModellingDemo.hill_curve — Method

hill_curve(x, half_max_concentration_point, hill_coef,::Val{:safe})

Hill Curve-based spend saturation. Safe implementation that works better with ForwardDiff

MediaMixModellingDemo.hill_curve — Method

hill_curve(x, half_max_concentration_point, hill_coef)

Hill Curve-based spend saturation based on: https://www.physiologyweb.com/calculators/hillequationinteractive_graph.html

MediaMixModellingDemo.mean_fitted_effects — Method

extract_fitted_effects(::Val{:model_stage2a},generated_quant,
    extract_keys=[:mu_trend,:mu_org,:mu_context,:mu_spend_by_var],mask=nothing)

Extract specific keys from the generated_quantities produced by Turing generated_quantities() and concatenates them It should be implemented for each model to ensure the right logic

Returns: a Vector (!)

Example: mean_fitted_effects(Val(:model_stage2a),stage2_fit_allsamples) mean_fitted_effects(Val(:model_stage2a),stage2_fit_allsamples;extract_keys=[:mu_spend_by_var])

MediaMixModellingDemo.multiply_and_normalize — Method

multiply_and_normalize(v::AbstractVector, multipliers::AbstractVector)

Multiplies a vector v (of adspend) by multipliers in a way that doesn't change the overall sum

MediaMixModellingDemo.plot_periodogram — Function

function plot_periodogram(input_arr,top_k)

Plot Fourier transform coefficients to uncover the most prominent frequencies / seasonalities Assumes equally spaced data points Looks only for periods that have seen at least 2 full cycles (ie, size ÷ 2 at maximum!) Shows top-k values

Example

p=10 # period is 10 y=sin.(2π/p*collect(1:20)) # generate 20 data points plot_periodogram(y,1) # plot periodogram, period=10 should be highlighted as maximum

MediaMixModellingDemo.plot_priors_decay_rate — Function

plot_priors_decay_rate(p,cols=nothing)

Plots priors for decay rate in terms of Beta distributions

Example:

p2=ParamsStage2()
cols_spend=nothing # available from data prep
plot_priors_decay_rate(p2,cols_spend)

MediaMixModellingDemo.saturate_adspend — Method

saturate_adspend(x::Number, chain::AbstractMCMC.AbstractChains, factor_to_roas_of_one::AbstractVector)

Extracts Hill Curve parameters and the corresponding beta coefficients from provided chain and applies them to a provided point (x)

WARNING: Depends on model implementation using the same RVs as modelstage2a, ie, `betaspend,halfmaxpoint,slope`

MediaMixModellingDemo.set_priors_auto_scales — Function

set_priors_auto_scales(y::AbstractArray,X,param_key::Symbol,factor::Union{Number,AbstractArray}=1.0,
    existing_params=ParamsStage1())

Sets scales such that 1 standard dev of input data (X) could trigger at most 3standard dev. response of Y (y) (because beta coefficient are Normal). This can be increased by factor (defaults to 1.0x). Result is saved into a Parameters object (as per constructor) under a key param_key It can extend (selectively update) existing set of parameters (provide to argument existing_params)

Special behaviour:

Values are clipped to range 0.05 - 10. (warning will be issued if it's exceeded)

Shortcuts to straight passthrough if X=nothing

MediaMixModellingDemo.set_priors_stage2_decay_rates — Function

set_priors_stage2_decay_rates(decay_rate_types::AbstractArray{String},
    decay_rates_types_dictionary::Dict=decay_rates_types_dictionary,
    existing_params=ParamsStage2(),var_names=nothing)

Sets decay rates priors (for Beta distribution) as a dictionary lookup of different options available in decay_rates_types_dictionary (eg, "tv", "digital")

Example:

types_=["ooh","digital","digital","ooh","ooh"]
set_priors_stage2_decay_rates(types_,decay_rates_types_dictionary,p2,cols_spend)

MediaMixModellingDemo.simulate_revenues_summed — Method

simulate_revenues_summed(chain_optim::AbstractMCMC.AbstractChains, model_optim::DynamicPPL.Model,

optimmask::BitVector;extractkey::Symbol=:y)

Samples posterior predictive from model model_optim conditioned on Chains chain_optim. Optional - you can provide a mask optim_mask (for time dimensions/1st dimension) and Symbol for the posterior predictive value (extract_key)

Example:

simulations_prev=simulate_revenues_summed(chain_optim,model_orig,optim_mask;extract_key=:y)

For budget simulations we replace inputs in the model like:

chain_optim=Chains(chain,:parameters)
model_args_prev=model_orig.args;
# replace the old spend with new
model_args_new=merge(model_args_prev,(;X_spend=X_spend_new));
simulations=simulate_revenues_summed(chain_optim,model_stage2a(model_args_new...),optim_mask);

MediaMixModellingDemo.standardize_by_max — Method

standardize_by_max(X)

Max()-only transform to allow easy scaling between features and the outcome Uses MinMax() pipe under the hood but overwrites the minimum to be =0

Example: y_std,pipe_cache_y=standardize_by_max(select(df,target_label))

MediaMixModellingDemo.standardize_by_zscore — Method

standardize_by_zscore(X)

Zscore transform to center the feature to its mean value and scale it (to make it easier to set priors)

Example: y_std,pipe_cache_y=standardize_by_zscore(select(df,target_label))

MediaMixModellingDemo.sum_columns — Method

sum_columns(x::AbstractMatrix)

Unified interface that sums columns of a provided Matrix/DataFrame/Vector Returns a Vector (!)

MediaMixModellingDemo.threaded_objective_func — Method

threaded_objective_func(budget_multipliers,objective_func)

Parallelizes provided function objective_func for Metaheuristics optimization loop by leveraging available threads in your Julia instance

Example:

Make sure to change the keyword to parallel_evaluation=true

options = Metaheuristics.Options(time_limit=10.,debug=false,parallel_evaluation=true)
@time result = Metaheuristics.optimize(x->threaded_objective_func(x,objective_func), bounds, 
    Metaheuristics.ECA(N=7*2*length(cols_spend),K=7,η_max=2.,options=options))

MediaMixModellingDemo.to_masked_matrix — Function

to_masked_matrix(x::DataFrame,mask=trues(size(x,1)))

Convert DataFrame to a matrix and apply a mask to its rows if provided

MediaMixModellingDemo.workflow_budget_to_simulation — Method

workflow_budget_to_simulation(
                chain_optim::AbstractMCMC.AbstractChains, model_orig::DynamicPPL.Model,
               X_spend::AbstractMatrix, optim_mask::BitVector,
               factor_to_scale_spend_to_orig::AbstractVector,
               budget_multiplier::AbstractVector)

For a given context (budget_multiplier) produces new adspend (X_spend_new) and associated simulated revenues (smulations_new) Used to produced an objective function for the optimization that depends only on budget_multiplier